requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
APScheduler==3.10.4
asyncpg==0.29.0
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
frozenlist==1.4.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
gspread==6.1.3
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
oauth2client==4.1.3
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
python-decouple==3.8
python-telegram-bot==21.6
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
yarl==1.13.1
